\name{utinit.nc}

\alias{utinit.nc}

\title{Initialize the UDUNITS Library}

\description{Initialize the UDUNITS library.}

\usage{utinit.nc(path="")}

\arguments{
  \item{path}{Path to a units file containing initializing unit definitions.}
}

\details{In previous versions of RNetCDF, this function initialized the UDUNITS library. It is longer used, and is provided purely for backward compatibility.
}

\references{\url{http://www.unidata.ucar.edu/software/udunits/}}

\author{Pavel Michna}

\keyword{utilities}
